Update time types for newer kernels
<Author: jwiele@redhat.com>

Newer kernels have removed time types and functions that are
specific to user space APIs as part of the effort to use
ktime_t for all kernel timekeeping. Luckily, UDS only uses
timespec and timeval in user space. UDS uses time_t in stats,
but time_t is just a number of seconds, and in modern systems
should be a 64 bit integer, even on 32 bit systems, so it can
be compatibly replaced with int64_t. A static assertion will
check that assumption.

Pair: sweettea

.../timeUtils.h
Add a function to convert absolute nanosecond time to
  milliseconds.
Wrap user space specific functions with "#ifndef __KERNEL__".
Correct the comments on the user space specific functions.

.../timeUtils.c
Add a static assertion that time_t is the same as int64_t.

.../indexLayout.c
Use the new absolute time to milliseconds function in lieu of
  the old getTimeMS function which was needed when user space
  time had a different representation from kernel space time,
  but can now be replaced with a simpler function.

.../uds.h
.../indexSession.c
Represent time in stats as int64_t instead of time_t.