Eliminated obsolete function smp_read_barrier_depends.
<Author: jwiele@redhat.com>

In Kernels newer than 5.9 or so, smp_read_barrier_depends has
been removed because its functionality is provided by
READ_ONCE.  See: https://lkml.org/lkml/2019/11/8/1021

We provide a user space implementation that is empty for the
four primary platforms we test on. That can be deleted. Some
additional barriers can be added if we find that some platform
requires them.

Pair: sweeettea

.../funnelQueue.c
Replaced the call to smp_read_barrier_depends with a comment
  noting that an additional barrier may be needed in the
  future for some platforms such as Alpha.